Seperate MIDISequencer classes
