+++ MIDISequencer.podspec
@@ -176,7 +176,7 @@ velocity: .standard(60))
   #  Not including the public_header_files will make all headers public.
   #
 
-  s.source_files  = "Source/MIDISequencer.swift"
+  s.source_files  = "Source/*.swift"
   # s.exclude_files = "Classes/Exclude"
 
   # s.public_header_files = "Classes/**/*.h"
